backend/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.7
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
alembic>=1.13.2
